proxy header to s3 igv reference requests
